Minor Bug Fixes and a Demo Matlab Script
Add example.m as a tutorial to using the function and I think its
pretty solid, but would like feedback.

The cutoff parameter was being dumb and now its not.

+++ Codes/FourierPad.m
@@ -1,7 +1,7 @@
 function fA = FourierPad( A, osz, nsz );
 
 if all( osz == nsz )
-    fA = A;
+    fA = double(A);
 else
     switch numel(nsz)
         case 1
